Adds KnownType Support to SmartEnum
For types that derive from`SmartEnum` directly to be properly serialized
as `SmartEnum`, the serializer must know that the derived type is a
`SmartEnum`. This is typically done by manually noting the derived types
if they are known ahead of time or dynamically building the list via
reflection if they are not.

In this case however, since `SmartEnum` is used only as a generic the
types that directly derive from `SmartEnum` can be returned directly as
via the type of `TEnum`.